fix for remove condition
