+++ include/aquarius/concepts.hpp
@@ -1,9 +1,10 @@
 #pragma once
 #include <aquarius/serialize/scalar_type.hpp>
 #include <concepts>
+#include <iterator>
+#include <map>
 #include <string>
 #include <string_view>
-#include <map>
 
 namespace aquarius
 {
